qwen-llm/lora_qlora/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Data processing
pandas>=1.3.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional dependencies for advanced features
accelerate>=0.20.0
bitsandbytes>=0.41.0
peft>=0.4.0

# Development dependencies
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
